Release v0.11.4 (#313)
* Added Password Prompt to operate with echo off in terminal
([#265](https://github.com/databrickslabs/blueprint/issues/265)). The
command-line interface now includes a secure password prompt feature,
allowing users to enter sensitive information without it being visible
on the screen. This is achieved through a new method that utilizes the
`getpass` library to hide user input when entering passwords, taking a
prompt message and an optional maximum number of attempts as parameters.
The method repeatedly prompts the user for a password until a valid
input is provided or the maximum number of attempts is reached, at which
point it raises a `ValueError`. This addition enhances the security and
usability of the interface, and its functionality is validated through
new test methods that cover both successful password entry and the
scenario where the maximum number of attempts is exceeded, ensuring the
feature behaves as expected in various situations.
* Sniff encoding properly in XML files with a standalone directive
([#256](https://github.com/databrickslabs/blueprint/issues/256)). The
XML file encoding detection has been improved to support a wider range
of valid XML declarations. The regular expression used to match XML
declarations has been updated to correctly handle cases where both
`encoding` and `standalone` attributes are present, such as `<?xml
version="1.x" encoding="xxx" standalone="yes"?>`. This change enables
more accurate detection of the encoding attribute in XML files, even
when a `standalone` directive is present. Additionally, test functions
have been added and modified to verify this functionality, including
tests for XML files with a BOM prefix and those with an XML standalone
declaration, ensuring that the code can correctly read these files and
detect the encoding.