Handle `TRACE`-level logging (#318)
This PR updates the logging set up during handoff from the Databricks
CLI so that `TRACE`-level logging doesn't trigger a crash. Because
Python doesn't support `TRACE` as a logging level we map this to
`DEBUG`.

In addition, rather than crash when an unknown log-level is encountered
we now log a warning and proceed with `INFO` instead.

Resolves #317.

This has been tested manually with `lakebridge`.