Project linting: support for Python 3.14 and general cleanup (#319)
This PR upgrades the linting tools for the project: the previous
versions did not work (and crashed) on Python 3.14 due to changes in the
underlying bytecode format. Changes include:

 - Upgrade mypy and pylint to their current versions.
 - Fixing some type-annotation issues due to the linters being stricter.
- Updating some of the method signatures that changed in Python 3.13 but
hadn't been updated, to keep the warnings from pylint across the
different versions to a minimum.

None of the changes affect backward-compatibility with respect to using
this package.